daudcut/requirements.txt
youtube-dl
discord
ffmpeg-python
python-libmagic

python-magic-bin; platform_system == "Windows"
